pyproject.toml
[tool.poetry]
name = "una-coding-task"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.6"
djangorestframework = "^3.13.1"
Markdown = "^3.3.7"
django-filter = "^22.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
